init: add fluxbind code for processor binding
This library will allow for specifying a shapefile (yaml)
that describes how processes should be bound. The user can
define everything from numa nodes to cores, l3 caches,
or basically anything that hwloc understands. This is a wip,
and next step is to test different shapes on actual hpc apps
across multiple nodes. If the results are different / interesting,
we can define this as part of descriptive expression for jobs.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>